+++ app/pricing_map_by_sku.go
@@ -10,18 +10,18 @@ import (
 	"github.com/go-chi/chi"
 )
 
-// ListPricingBySKUHandler creates a handler function that returns a
-// map of tier refs to pricing entries.
-func (a *App) ListPricingBySKUHandler() http.HandlerFunc {
+// PricingMapBySKUHandler creates a handler function that returns a
+// map of tier refs to PricingEntry.
+func (a *App) PricingMapBySKUHandler() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		sku := chi.URLParam(r, "sku")
-		pmap, err := a.Service.ListPricingBySKU(r.Context(), sku)
+		pmap, err := a.Service.PricingMapBySKU(r.Context(), sku)
 		if err != nil {
 			if err == sql.ErrNoRows {
 				w.WriteHeader(http.StatusNotFound)
 				return
 			}
-			fmt.Fprintf(os.Stderr, "service ListPricingBySKU(ctx, %s) error: %+v", sku, err)
+			fmt.Fprintf(os.Stderr, "service PricingMapBySKU(ctx, %s) error: %+v", sku, err)
 			w.WriteHeader(http.StatusInternalServerError) // 500 Internal Server Error
 			return
 		}